Fix sonar
